Isolate and rename types
