+++ packages/twenty-front/src/modules/spreadsheet-import/steps/components/UploadStep/components/ExampleTable.tsx
@@ -1,13 +1,13 @@
 import { useMemo } from 'react';
 
 import { SpreadsheetImportTable } from '@/spreadsheet-import/components/SpreadsheetImportTable';
-import { Fields } from '@/spreadsheet-import/types';
+import { SpreadsheetImportFields } from '@/spreadsheet-import/types';
 import { generateExampleRow } from '@/spreadsheet-import/utils/generateExampleRow';
 
 import { generateColumns } from './columns';
 
 interface ExampleTableProps<T extends string> {
-  fields: Fields<T>;
+  fields: SpreadsheetImportFields<T>;
 }
 
 export const ExampleTable = <T extends string>({
