+++ packages/twenty-front/src/modules/spreadsheet-import/steps/components/ValidationStep/types.ts
@@ -1,8 +1,8 @@
-import { Info } from '@/spreadsheet-import/types';
+import { SpreadsheetImportInfo } from '@/spreadsheet-import/types';
 
 export type ImportedStructuredRowMetadata = {
   __index: string;
   __errors?: Error | null;
 };
-export type Error = { [key: string]: Info };
+export type Error = { [key: string]: SpreadsheetImportInfo };
 export type Errors = { [id: string]: Error };
