+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportErrorLevel.ts
@@ -1,0 +1,1 @@
+export type SpreadsheetImportErrorLevel = 'info' | 'warning' | 'error';
