+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportFields.ts
@@ -1,0 +1,6 @@
+import { SpreadsheetImportField } from '@/spreadsheet-import/types/SpreadsheetImportField';
+import { ReadonlyDeep } from 'type-fest';
+
+export type SpreadsheetImportFields<T extends string> = ReadonlyDeep<
+  SpreadsheetImportField<T>[]
+>;
