+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportImportValidationResult.ts
@@ -1,0 +1,9 @@
+import { ImportedStructuredRowMetadata } from '@/spreadsheet-import/steps/components/ValidationStep/types';
+import { ImportedStructuredRow } from './SpreadsheetImportImportedStructuredRow';
+
+export type SpreadsheetImportImportValidationResult<T extends string> = {
+  validStructuredRows: ImportedStructuredRow<T>[];
+  invalidStructuredRows: ImportedStructuredRow<T>[];
+  allStructuredRows: (ImportedStructuredRow<T> &
+    ImportedStructuredRowMetadata)[];
+};
