+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportImportedRow.ts
@@ -1,0 +1,1 @@
+export type ImportedRow = Array<string | undefined>;
