+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportImportedStructuredRow.ts
@@ -1,0 +1,3 @@
+export type ImportedStructuredRow<T extends string> = {
+  [key in T]: string | boolean | undefined;
+};
