+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportInfo.ts
@@ -1,0 +1,6 @@
+import { SpreadsheetImportErrorLevel } from './SpreadsheetImportErrorLevel';
+
+export type SpreadsheetImportInfo = {
+  message: string;
+  level: SpreadsheetImportErrorLevel;
+};
