+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportRowHook.ts
@@ -1,0 +1,8 @@
+import { ImportedStructuredRow } from './SpreadsheetImportImportedStructuredRow';
+import { SpreadsheetImportInfo } from './SpreadsheetImportInfo';
+
+export type SpreadsheetImportRowHook<T extends string> = (
+  row: ImportedStructuredRow<T>,
+  addError: (fieldKey: T, error: SpreadsheetImportInfo) => void,
+  table: ImportedStructuredRow<T>[],
+) => ImportedStructuredRow<T>;
