+++ packages/twenty-front/src/modules/spreadsheet-import/types/SpreadsheetImportTableHook.ts
@@ -1,0 +1,11 @@
+import { ImportedStructuredRow } from './SpreadsheetImportImportedStructuredRow';
+import { SpreadsheetImportInfo } from './SpreadsheetImportInfo';
+
+export type SpreadsheetImportTableHook<T extends string> = (
+  table: ImportedStructuredRow<T>[],
+  addError: (
+    rowIndex: number,
+    fieldKey: T,
+    error: SpreadsheetImportInfo,
+  ) => void,
+) => ImportedStructuredRow<T>[];
