+++ packages/twenty-front/src/modules/spreadsheet-import/utils/findMatch.ts
@@ -1,6 +1,5 @@
+import { SpreadsheetImportFields } from '@/spreadsheet-import/types';
 import lavenstein from 'js-levenshtein';
-
-import { Fields } from '@/spreadsheet-import/types';
 
 type AutoMatchAccumulator<T> = {
   distance: number;
@@ -9,7 +8,7 @@ type AutoMatchAccumulator<T> = {
 
 export const findMatch = <T extends string>(
   header: string,
-  fields: Fields<T>,
+  fields: SpreadsheetImportFields<T>,
   autoMapDistance: number,
 ): T | undefined => {
   const smallestValue = fields.reduce<AutoMatchAccumulator<T>>((acc, field) => {
