+++ packages/twenty-front/src/modules/spreadsheet-import/utils/generateExampleRow.ts
@@ -1,13 +1,21 @@
-import { Field, Fields } from '@/spreadsheet-import/types';
+import {
+  SpreadsheetImportField,
+  SpreadsheetImportFields,
+} from '@/spreadsheet-import/types';
 
-const titleMap: Record<Field<string>['fieldType']['type'], string> = {
+const titleMap: Record<
+  SpreadsheetImportField<string>['fieldType']['type'],
+  string
+> = {
   checkbox: 'Boolean',
   select: 'Options',
   multiSelect: 'Options',
   input: 'Text',
 };
 
-export const generateExampleRow = <T extends string>(fields: Fields<T>) => [
+export const generateExampleRow = <T extends string>(
+  fields: SpreadsheetImportFields<T>,
+) => [
   fields.reduce(
     (acc, field) => {
       acc[field.key as T] = field.example || titleMap[field.fieldType.type];
