+++ packages/twenty-front/src/modules/ui/field/input/components/MultiSelectInput.tsx
@@ -13,9 +13,9 @@ import { useSelectableListStates } from '@/ui/layout/selectable-list/hooks/inter
 import { useSelectableList } from '@/ui/layout/selectable-list/hooks/useSelectableList';
 import { useScopedHotkeys } from '@/ui/utilities/hotkey/hooks/useScopedHotkeys';
 import { useListenClickOutside } from '@/ui/utilities/pointer-event/hooks/useListenClickOutside';
+import { isDefined } from 'twenty-shared/utils';
 import { MenuItemMultiSelectTag } from 'twenty-ui';
 import { turnIntoEmptyStringIfWhitespacesOnly } from '~/utils/string/turnIntoEmptyStringIfWhitespacesOnly';
-import { isDefined } from 'twenty-shared/utils';
 
 type MultiSelectInputProps = {
   selectableListComponentInstanceId: string;
